Refactor Dialog Folder Setup
