+++ frontend/src/app/my-courses/my-courses.module.ts
@@ -41,7 +41,7 @@ import { CalledTicketCardWidget } from './course/office-hours/widgets/called-tic
 import { QueuedTicketCardWidget } from './course/office-hours/widgets/queued-ticket-card/queued-ticket-card.widget';
 import { OfficeHoursGetHelpComponent } from './course/office-hours/office-hours-get-help/office-hours-get-help.component';
 import { MatInputModule } from '@angular/material/input';
-import { CreateCourseSiteDialog } from './my-courses-page/dialogs/create-course-site.dialog';
+import { CreateCourseSiteDialog } from './dialogs/create-course-site/create-course-site.dialog';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { SettingsComponent } from './course/settings/settings.component';
